:recycle: refactoring: Update deletarJogo method to throw RuntimeException for not found Jogo instead of using ResponseStatusException

+++ src/main/java/com/conectados/conectados/services/JogoService.java
@@ -36,6 +36,10 @@ public class JogoService {
     }
 
     public void deletarJogo(Long id) {
+        Optional<Jogo> jogo = buscarJogoPorId(id);
+        if (jogo.isEmpty()) {
+            throw new RuntimeException("Jogo não encontrado");
+        }
         jogoRepository.deleteById(id);
     }
